refactor(core): consolidate serialized JSON byte counting
Three hand-written counters answered the same question — how many UTF-8
bytes a value's JSON representation occupies. @maka/code-mode carried a
bounded, early-exit implementation; additional-permissions.ts and
sandbox-boundary.ts each carried a byte-identical JSON.stringify wrapper.

Move the bounded implementation to @maka/core/serialized-byte-length and
point all three call sites at it. The two core call sites pass freshly
built, structurally validated payloads — plain objects, arrays, strings
and booleans, with no undefined, function, toJSON hook or cycle — for
which the bounded counter and JSON.stringify agree byte for byte,
including escapes, control characters, lone surrogates and multi-byte
sequences. A test pins that agreement.

Generated-by: Claude Code
